move code to search_structural.go
