+++ cmd/frontend/graphqlbackend/search_structural_test.go
@@ -3,9 +3,12 @@ package graphqlbackend
 import (
 	"context"
 	"errors"
+	"reflect"
 	"sync"
 	"testing"
 	"time"
+
+	"github.com/hashicorp/go-multierror"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/zoekt"
@@ -127,4 +130,47 @@ func TestBuildQuery(t *testing.T) {
 			t.Error(diff)
 		}
 	})
+}
+
+func TestConvertErrorsForStructuralSearch(t *testing.T) {
+	cases := []struct {
+		name       string
+		errors     []error
+		wantErrors []error
+	}{
+		{
+			name:       "multierr_is_unaffected",
+			errors:     []error{errors.New("some error")},
+			wantErrors: []error{errors.New("some error")},
+		},
+		{
+			name: "convert_text_errors_to_typed_errors",
+			errors: []error{
+				errors.New("some error"),
+				errors.New("Worker_oomed"),
+				errors.New("some other error"),
+				errors.New("Out of memory"),
+				errors.New("yet another error"),
+				errors.New("no indexed repositories for structural search"),
+			},
+			wantErrors: []error{
+				errors.New("some error"),
+				errStructuralSearchMem,
+				errors.New("some other error"),
+				errStructuralSearchSearcher,
+				errors.New("yet another error"),
+				errStructuralSearchNoIndexedRepos{msg: "Learn more about managing indexed repositories in our documentation: https://docs.sourcegraph.com/admin/search#indexed-search."},
+			},
+		},
+	}
+	for _, test := range cases {
+		multiErr := &multierror.Error{
+			Errors:      test.errors,
+			ErrorFormat: multierror.ListFormatFunc,
+		}
+		haveMultiErr := convertErrorsForStructuralSearch(multiErr)
+		if !reflect.DeepEqual(haveMultiErr.Errors, test.wantErrors) {
+			t.Fatalf("test %s, have errors: %q, want: %q", test.name, haveMultiErr.Errors, test.wantErrors)
+		}
+	}
 }
